Seperation into g file
